chore: remove dead code and a redundant conversion
Tier-1 dead code (zero callers; verified via deadcode + grep):
- NewSimulatorManager (thin wrapper, superseded by NewSimulatorManagerWithOptions)
- SimulatorManager.IsUsingNamespace, SimulatorManager.GetNamespaceName

Tier-2 nit:
- gnmi_paths.go: drop redundant uint8() on OperUp/AdminUp (already uint8).

Deliberately NOT removed: encodeNull and Topology.AddLink (deadcode flagged
them but both are test-used). SetupRoutesForDevices is left in place — removing
it would orphan the Linux-only netns.AddRouteForDevices, which can't be
verified on this (darwin) build; tracked as a Linux-only follow-up. The
staticcheck quickfix hints (if/else->tagged switch in SNMP/ssh parse paths,
WriteString->Fprintf) are pure style in sensitive paths and were left out.

+++ go/nl6/gnmi_paths.go
@@ -469,14 +469,14 @@ func (r *pathResolver) resolveLeaf(ic *IfCounterCycler, ifIndex int, leaf string
 		// Read from the interface state engine (post add-interface-state).
 		// nil-state defensive path returns the OpenConfig UP default —
 		// matches behavior pre-state-engine for backward-compat.
-		v := uint8(OperUp)
+		v := OperUp
 		if state := ic.State(); state != nil {
 			v = state.OperStatus(ifIndex)
 		}
 		value = operStatusOpenConfig(v)
 	case "admin-status":
 		path = buildStateLeafPath(ifName, "admin-status")
-		v := uint8(AdminUp)
+		v := AdminUp
 		if state := ic.State(); state != nil {
 			v = state.AdminStatus(ifIndex)
 		}
